Move media items into dedicated files
