+++ src/media/episode.rs
@@ -1,0 +1,221 @@
+use crate::common::Image;
+use crate::crunchyroll::Executor;
+use crate::media::util::parse_locale_from_slug_title;
+use crate::media::{request_media, Media};
+use crate::{Crunchyroll, Locale, Result, Season, Series};
+use chrono::{DateTime, Duration, Utc};
+use serde::Deserialize;
+use std::sync::Arc;
+
+#[allow(dead_code)]
+#[derive(Clone, Debug, Default, Deserialize)]
+#[cfg_attr(feature = "__test_strict", serde(deny_unknown_fields))]
+#[cfg_attr(not(feature = "__test_strict"), serde(default))]
+pub(crate) struct EpisodeVersion {
+    #[serde(rename = "guid")]
+    pub(crate) id: String,
+    #[serde(rename = "media_guid")]
+    pub(crate) media_id: String,
+    #[serde(rename = "season_guid")]
+    pub(crate) season_id: String,
+
+    pub(crate) audio_locale: Locale,
+
+    pub(crate) is_premium_only: bool,
+    pub(crate) original: bool,
+
+    pub(crate) variant: String,
+}
+
+/// Metadata for a episode.
+#[allow(dead_code)]
+#[derive(Clone, Debug, Deserialize, smart_default::SmartDefault)]
+#[serde(remote = "Self")]
+#[cfg_attr(feature = "__test_strict", serde(deny_unknown_fields))]
+#[cfg_attr(not(feature = "__test_strict"), serde(default))]
+pub struct Episode {
+    #[serde(skip)]
+    pub(crate) executor: Arc<Executor>,
+
+    pub id: String,
+    #[serde(alias = "streams_link")]
+    #[serde(deserialize_with = "crate::internal::serde::deserialize_streams_link")]
+    pub stream_id: String,
+    pub channel_id: String,
+    #[serde(alias = "playback")]
+    pub playback_url: String,
+    pub identifier: String,
+
+    pub slug: String,
+    pub title: String,
+    pub slug_title: String,
+    pub description: String,
+
+    // both missing if the episode is the last one in its season unpopulated
+    #[serde(default)]
+    pub next_episode_id: String,
+    #[serde(default)]
+    pub next_episode_title: String,
+
+    pub season_id: String,
+    pub season_title: String,
+    pub season_slug_title: String,
+    #[serde(default)]
+    pub season_tags: Vec<String>,
+
+    pub series_id: String,
+    pub series_title: String,
+    pub series_slug_title: String,
+
+    // probably empty
+    #[serde(default)]
+    pub production_episode_id: String,
+
+    /// Usually the same as [`Episode::episode_number`], just as string.
+    pub episode: String,
+    #[serde(deserialize_with = "crate::internal::serde::deserialize_maybe_null_to_default")]
+    pub episode_number: u32,
+    /// Usually also the same as [`Episode::episode_number`]. If the episode number is null (which
+    /// occurs for the first AOT episode, which is a preview, for example) this might be a floating
+    /// number like 0.5.
+    pub sequence_number: f32,
+
+    pub season_number: u32,
+
+    pub audio_locale: Locale,
+    pub subtitle_locales: Vec<Locale>,
+
+    #[serde(alias = "duration_ms")]
+    #[serde(deserialize_with = "crate::internal::serde::deserialize_millis_to_duration")]
+    #[default(Duration::milliseconds(0))]
+    pub duration: Duration,
+
+    #[default(DateTime::<Utc>::from(std::time::SystemTime::UNIX_EPOCH))]
+    pub episode_air_date: DateTime<Utc>,
+    /// The same as episode_air_date as far as I can see.
+    #[default(DateTime::<Utc>::from(std::time::SystemTime::UNIX_EPOCH))]
+    pub upload_date: DateTime<Utc>,
+    #[default(DateTime::<Utc>::from(std::time::SystemTime::UNIX_EPOCH))]
+    pub free_available_date: DateTime<Utc>,
+    #[default(DateTime::<Utc>::from(std::time::SystemTime::UNIX_EPOCH))]
+    pub premium_available_date: DateTime<Utc>,
+    #[default(DateTime::<Utc>::from(std::time::SystemTime::UNIX_EPOCH))]
+    pub availability_starts: DateTime<Utc>,
+    #[default(DateTime::<Utc>::from(std::time::SystemTime::UNIX_EPOCH))]
+    pub availability_ends: DateTime<Utc>,
+
+    #[serde(deserialize_with = "crate::internal::serde::deserialize_thumbnail_image")]
+    pub images: Vec<Image>,
+
+    pub is_dubbed: bool,
+    pub is_subbed: bool,
+
+    pub is_premium_only: bool,
+    pub is_clip: bool,
+
+    pub is_mature: bool,
+    pub maturity_ratings: Vec<String>,
+    pub mature_blocked: bool,
+
+    pub available_offline: bool,
+    pub availability_notes: String,
+
+    pub closed_captions_available: bool,
+
+    pub eligible_region: String,
+
+    #[serde(default)]
+    pub(crate) versions: Option<Vec<EpisodeVersion>>,
+
+    #[cfg(feature = "__test_strict")]
+    media_type: Option<crate::StrictValue>,
+    #[cfg(feature = "__test_strict")]
+    external_id: Option<crate::StrictValue>,
+    #[cfg(feature = "__test_strict")]
+    linked_resource_key: Option<crate::StrictValue>,
+    #[cfg(feature = "__test_strict")]
+    new: Option<crate::StrictValue>,
+    #[cfg(feature = "__test_strict")]
+    promo_title: Option<crate::StrictValue>,
+    #[cfg(feature = "__test_strict")]
+    promo_description: Option<crate::StrictValue>,
+    #[cfg(feature = "__test_strict")]
+    search_metadata: Option<crate::StrictValue>,
+    #[cfg(feature = "__test_strict")]
+    #[serde(rename = "type")]
+    _type: Option<crate::StrictValue>,
+    #[cfg(feature = "__test_strict")]
+    extended_maturity_rating: crate::StrictValue,
+    #[cfg(feature = "__test_strict")]
+    tenant_categories: Option<crate::StrictValue>,
+    #[cfg(feature = "__test_strict")]
+    available_date: crate::StrictValue,
+    #[cfg(feature = "__test_strict")]
+    premium_date: crate::StrictValue,
+    #[cfg(feature = "__test_strict")]
+    seo_title: Option<crate::StrictValue>,
+    #[cfg(feature = "__test_strict")]
+    seo_description: Option<crate::StrictValue>,
+    #[cfg(feature = "__test_strict")]
+    listing_id: Option<crate::StrictValue>,
+    #[cfg(feature = "__test_strict")]
+    hd_flag: Option<crate::StrictValue>,
+}
+
+impl Episode {
+    /// Returns the series the episode belongs to.
+    pub async fn series(&self) -> Result<Series> {
+        let endpoint = format!(
+            "https://www.crunchyroll.com/content/v2/cms/series/{}",
+            self.series_id
+        );
+        Ok(request_media(self.executor.clone(), endpoint)
+            .await?
+            .remove(0))
+    }
+
+    /// Returns the season the episode belongs to.
+    pub async fn season(&self) -> Result<Season> {
+        let endpoint = format!(
+            "https://www.crunchyroll.com/content/v2/cms/seasons/{}",
+            self.season_id
+        );
+        Ok(request_media(self.executor.clone(), endpoint)
+            .await?
+            .remove(0))
+    }
+}
+
+#[async_trait::async_trait]
+impl Media for Episode {
+    async fn from_id(crunchyroll: &Crunchyroll, id: impl AsRef<str> + Send) -> Result<Self> {
+        Ok(request_media(
+            crunchyroll.executor.clone(),
+            format!(
+                "https://www.crunchyroll.com/content/v2/cms/episodes/{}",
+                id.as_ref()
+            ),
+        )
+        .await?
+        .remove(0))
+    }
+
+    #[cfg(feature = "experimental-stabilizations")]
+    async fn __apply_experimental_stabilizations(&mut self) {
+        if self.executor.fixes.locale_name_parsing {
+            self.audio_locale = parse_locale_from_slug_title(&self.season_slug_title)
+        }
+        if self.executor.fixes.season_number {
+            let mut split = self.identifier.splitn(3, '|');
+            let (_, season, _) = (
+                split.next().unwrap_or_default(),
+                split.next().unwrap_or_default(),
+                split.next().unwrap_or_default(),
+            );
+
+            if let Ok(season_num) = season.trim_start_matches('S').parse() {
+                self.season_number = season_num
+            }
+        }
+    }
+}
